requirements.txt
pandas
numpy==1.21.0
seaborn
matplotlib
nltk
tensorflow==2.9.3
keras
regex
scikit-learn
fastapi==0.78.0
python-dotenv
pymongo
scikit-learn
-e .